Refactor FileResultInfo to FileResultInfoDto
