+++ src/main/java/com/neconico/neconico/dto/file/FileResultInfoDto.java
@@ -1,14 +1,16 @@
-package com.neconico.neconico.immutable;
+package com.neconico.neconico.dto.file;
 
 import lombok.Getter;
+import lombok.Setter;
 
+@Setter
 @Getter
-public class FileResultInfo {
+public class FileResultInfoDto {
     private String fileUrls;
 
     private String fileNames;
 
-    public FileResultInfo(String fileUrls, String fileNames) {
+    public FileResultInfoDto(String fileUrls, String fileNames) {
         this.fileUrls = fileUrls;
         this.fileNames = fileNames;
     }
