+++ src/main/java/com/neconico/neconico/file/process/FileProcess.java
@@ -1,13 +1,13 @@
 package com.neconico.neconico.file.process;
 
-import com.neconico.neconico.immutable.FileResultInfo;
+import com.neconico.neconico.dto.file.FileResultInfoDto;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 
 public interface FileProcess {
 
-    FileResultInfo uploadFile(MultipartFile... files) throws IOException, IllegalStateException, IllegalArgumentException;
+    FileResultInfoDto uploadFile(MultipartFile... files) throws IOException, IllegalStateException, IllegalArgumentException;
 
     boolean canDeleteFiles(String fileNames) throws IllegalArgumentException;
