+++ src/main/java/com/neconico/neconico/service/file/FileService.java
@@ -1,7 +1,7 @@
 package com.neconico.neconico.service.file;
 
 import com.neconico.neconico.file.process.FileProcess;
-import com.neconico.neconico.immutable.FileResultInfo;
+import com.neconico.neconico.dto.file.FileResultInfoDto;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
@@ -10,7 +10,7 @@ public interface FileService {
 
     void setFileProcess(FileProcess fileProcess);
 
-    FileResultInfo uploadFiles(MultipartFile... files) throws IOException, IllegalStateException, IllegalArgumentException;
+    FileResultInfoDto uploadFiles(MultipartFile... files) throws IOException, IllegalStateException, IllegalArgumentException;
 
     boolean canDeleteFiles(String fileNames) throws IllegalArgumentException;
